--- postwork6/postwork6/src/main/java/org/bedu/backend/postwork6/dao/PersonaDAO.java
@@ -1,22 +1,0 @@
-package org.bedu.backend.postwork6.dao;
-
-import org.bedu.backend.postwork6.model.Persona;
-import org.springframework.stereotype.Repository;
-
-import java.util.SortedSet;
-import java.util.TreeSet;
-
-@Repository
-public class PersonaDAO {
-
-    private static SortedSet<Persona> personas = new TreeSet<>();
-
-    public Persona guardaPersona(Persona persona) {
-        personas.add(persona);
-        return persona;
-    }
-
-    public SortedSet<Persona> getPersonas(){
-        return personas;
-    }
-}
